Prefab for Defense Troop
Created a prefab for a Defense Troop with a Collider2D component (Is Trigger) and added a script to locate targets, TargetRadar.

-------------------------------------------
Target Radar script

This script will locate the nearest target within a range set by the Collider2D component.

[SerializeField] string targetTag - Tag of the object to be targeted.

void OnTriggerEnter2D - Adds to the possibleTargets list all Collider2D that enters the trigger from this.gameObject if their tag is the same as targetTag.

void OnTriggerExit2D - Removes from the possibleTargets list all Collider2D that exits the trigger from this.gameObject.

void Update() - Searches for closest target every frame.

GameObject FindClosestTarget() - Returns the gameObject with the closest position to this.gameObject from the possibleTargets list.