Main menu scene
Created a scene for the Main Menu with two buttons: StartButton to load the main scene and QuitButton to close the game.

Added a script for the Canvas.MainMenu gameObject with a MainSceneLoader and QuitGame functions.

public void MainSceneLoader() - Loads the scene on the indexed location set on mainSceneIndex from the build.

public void QuitGame() - Closes the application.