Shopping system created
Changes to CurrentMoney and SpawnAreaScript scripts plus corrections on Panel and GenericArea placement.

Changes on CurrentMoney.cs:

variables:

_currentMoney is now a SerializedField, and sets the initial money on the phase.

Functions:

Erased void RefreshTheCurrentMoney(int) function and replaced on Update() where it alters the text according to the _currentMoney variable.

Created bool UpdateMoney(int) function where alters _currentMoney value as long as it doesn't get it under 0.

Changes on SpawnAreaScript.cs:

variables:

Added a SerializeField GameObject currentMoney. The gameObject that controls and display the current money on the game.

Functions:

Changes on OnMouseDown() where it places a character only when there is enough money avaiable.
TODO: Character cost on the buttons.